Code cleanup

+++ crypto/src/crypto/modes/GcmSivBlockCipher.cs
@@ -37,17 +37,17 @@ namespace Org.BouncyCastle.Crypto.Modes
         * The maximum data length (AEAD/PlainText). Due to implementation constraints this is restricted to the maximum
         * array length (https://programming.guide/java/array-maximum-length.html) minus the BUFLEN to allow for the MAC
         */
-        private static readonly int MAX_DATALEN = Int32.MaxValue - 8 - BUFLEN;
+        private static readonly int MAX_DATALEN = int.MaxValue - 8 - BUFLEN;
 
         /**
         * The top bit mask.
         */
-        private static readonly byte MASK = (byte)0x80;
+        private static readonly byte MASK = 0x80;
 
         /**
         * The addition constant.
         */
-        private static readonly byte ADD = (byte)0xE1;
+        private static readonly byte ADD = 0xE1;
 
         /**
         * The initialisation flag.
@@ -246,7 +246,7 @@ namespace Org.BouncyCastle.Crypto.Modes
             }
 
             /* Make sure that we haven't breached AEAD data limit */
-            if ((long)theAEADHasher.getBytesProcessed() + Int64.MinValue > (MAX_DATALEN - pLen) + Int64.MinValue)
+            if ((long)theAEADHasher.getBytesProcessed() + long.MinValue > (MAX_DATALEN - pLen) + long.MinValue)
             {
                 throw new InvalidOperationException("AEAD byte count exceeded");
             }
@@ -279,8 +279,7 @@ namespace Org.BouncyCastle.Crypto.Modes
                 dataLimit += BUFLEN;
                 currBytes = theEncData.Length;
             }
-            if (currBytes + System.Int64.MinValue
-            > (dataLimit - pLen) + System.Int64.MinValue)
+            if (currBytes + long.MinValue > (dataLimit - pLen) + long.MinValue)
             {
                 throw new InvalidOperationException("byte count exceeded");
             }
